Conflicts:
	flight/targets/FlyingF3/System/inc/pios_usb_board_data.h
	ground/gcs/src/libs/opmapcontrol/src/mapwidget/mapwidget.pro
	ground/gcs/src/libs/qscispinbox/qscispinbox.pro
	ground/gcs/src/libs/qwt/src/src.pro
	ground/gcs/src/plugins/config/defaulthwsettingswidget.h
	ground/gcs/src/plugins/config/hwfieldselector.h
	ground/gcs/src/plugins/coreplugin/coreplugin.pro
	ground/gcs/src/plugins/coreplugin/dialogs/settingsdialog.cpp
	ground/gcs/src/plugins/coreplugin/outputpane.h
	ground/gcs/src/plugins/debuggadget/debuggadget.pro
	ground/gcs/src/plugins/emptygadget/emptygadget.pro
	ground/gcs/src/plugins/hitl/plugin.pro
	ground/gcs/src/plugins/ipconnection/ipconnection.pro
	ground/gcs/src/plugins/notify/notify.pro
	ground/gcs/src/plugins/plugins.pro
	ground/gcs/src/plugins/rawhid/pjrc_rawhid_unix.cpp
	ground/gcs/src/plugins/rawhid/rawhid.pro
	ground/gcs/src/plugins/serialconnection/serialconnection.pro
	ground/gcs/src/plugins/uavobjectutil/uavobjectutil.pro
	ground/gcs/src/plugins/uavtalk/uavtalk.pro
	ground/gcs/src/plugins/welcome/welcome.pro
	make/tools.mk

+++ flight/targets/Bootloaders/DiscoveryF4/pios_board.c
@@ -37,7 +37,7 @@
 #include <pios_board_info.h>
 #include <pios.h>
 
-uint32_t pios_com_telem_usb_id;
+uintptr_t pios_com_telem_usb_id;
 
 static bool board_init_complete = false;
 void PIOS_Board_Init() {
